+++ api-macros/tests/expand/expand_02_state_clone.rs
@@ -1,0 +1,5 @@
+#[derive(Clone, trillium_api::TryFromConn)]
+#[api(state, clone)]
+struct CurrentUser {
+    name: String,
+}