+++ api-macros/tests/expand/expand_05_body.expanded.rs
@@ -1,0 +1,248 @@
+use serde::{Deserialize, Serialize};
+#[api(body)]
+struct Echo {
+    payload: String,
+}
+#[doc(hidden)]
+#[allow(
+    non_upper_case_globals,
+    unused_attributes,
+    unused_qualifications,
+    clippy::absolute_paths,
+)]
+const _: () = {
+    #[allow(unused_extern_crates, clippy::useless_attribute)]
+    extern crate serde as _serde;
+    #[automatically_derived]
+    impl _serde::Serialize for Echo {
+        fn serialize<__S>(
+            &self,
+            __serializer: __S,
+        ) -> _serde::__private228::Result<__S::Ok, __S::Error>
+        where
+            __S: _serde::Serializer,
+        {
+            let mut __serde_state = _serde::Serializer::serialize_struct(
+                __serializer,
+                "Echo",
+                false as usize + 1,
+            )?;
+            _serde::ser::SerializeStruct::serialize_field(
+                &mut __serde_state,
+                "payload",
+                &self.payload,
+            )?;
+            _serde::ser::SerializeStruct::end(__serde_state)
+        }
+    }
+};
+#[doc(hidden)]
+#[allow(
+    non_upper_case_globals,
+    unused_attributes,
+    unused_qualifications,
+    clippy::absolute_paths,
+)]
+const _: () = {
+    #[allow(unused_extern_crates, clippy::useless_attribute)]
+    extern crate serde as _serde;
+    #[automatically_derived]
+    impl<'de> _serde::Deserialize<'de> for Echo {
+        fn deserialize<__D>(
+            __deserializer: __D,
+        ) -> _serde::__private228::Result<Self, __D::Error>
+        where
+            __D: _serde::Deserializer<'de>,
+        {
+            #[allow(non_camel_case_types)]
+            #[doc(hidden)]
+            enum __Field {
+                __field0,
+                __ignore,
+            }
+            #[doc(hidden)]
+            struct __FieldVisitor;
+            #[automatically_derived]
+            impl<'de> _serde::de::Visitor<'de> for __FieldVisitor {
+                type Value = __Field;
+                fn expecting(
+                    &self,
+                    __formatter: &mut _serde::__private228::Formatter,
+                ) -> _serde::__private228::fmt::Result {
+                    _serde::__private228::Formatter::write_str(
+                        __formatter,
+                        "field identifier",
+                    )
+                }
+                fn visit_u64<__E>(
+                    self,
+                    __value: u64,
+                ) -> _serde::__private228::Result<Self::Value, __E>
+                where
+                    __E: _serde::de::Error,
+                {
+                    match __value {
+                        0u64 => _serde::__private228::Ok(__Field::__field0),
+                        _ => _serde::__private228::Ok(__Field::__ignore),
+                    }
+                }
+                fn visit_str<__E>(
+                    self,
+                    __value: &str,
+                ) -> _serde::__private228::Result<Self::Value, __E>
+                where
+                    __E: _serde::de::Error,
+                {
+                    match __value {
+                        "payload" => _serde::__private228::Ok(__Field::__field0),
+                        _ => _serde::__private228::Ok(__Field::__ignore),
+                    }
+                }
+                fn visit_bytes<__E>(
+                    self,
+                    __value: &[u8],
+                ) -> _serde::__private228::Result<Self::Value, __E>
+                where
+                    __E: _serde::de::Error,
+                {
+                    match __value {
+                        b"payload" => _serde::__private228::Ok(__Field::__field0),
+                        _ => _serde::__private228::Ok(__Field::__ignore),
+                    }
+                }
+            }
+            #[automatically_derived]
+            impl<'de> _serde::Deserialize<'de> for __Field {
+                #[inline]
+                fn deserialize<__D>(
+                    __deserializer: __D,
+                ) -> _serde::__private228::Result<Self, __D::Error>
+                where
+                    __D: _serde::Deserializer<'de>,
+                {
+                    _serde::Deserializer::deserialize_identifier(
+                        __deserializer,
+                        __FieldVisitor,
+                    )
+                }
+            }
+            #[doc(hidden)]
+            struct __Visitor<'de> {
+                marker: _serde::__private228::PhantomData<Echo>,
+                lifetime: _serde::__private228::PhantomData<&'de ()>,
+            }
+            #[automatically_derived]
+            impl<'de> _serde::de::Visitor<'de> for __Visitor<'de> {
+                type Value = Echo;
+                fn expecting(
+                    &self,
+                    __formatter: &mut _serde::__private228::Formatter,
+                ) -> _serde::__private228::fmt::Result {
+                    _serde::__private228::Formatter::write_str(
+                        __formatter,
+                        "struct Echo",
+                    )
+                }
+                #[inline]
+                fn visit_seq<__A>(
+                    self,
+                    mut __seq: __A,
+                ) -> _serde::__private228::Result<Self::Value, __A::Error>
+                where
+                    __A: _serde::de::SeqAccess<'de>,
+                {
+                    let __field0 = match _serde::de::SeqAccess::next_element::<
+                        String,
+                    >(&mut __seq)? {
+                        _serde::__private228::Some(__value) => __value,
+                        _serde::__private228::None => {
+                            return _serde::__private228::Err(
+                                _serde::de::Error::invalid_length(
+                                    0usize,
+                                    &"struct Echo with 1 element",
+                                ),
+                            );
+                        }
+                    };
+                    _serde::__private228::Ok(Echo { payload: __field0 })
+                }
+                #[inline]
+                fn visit_map<__A>(
+                    self,
+                    mut __map: __A,
+                ) -> _serde::__private228::Result<Self::Value, __A::Error>
+                where
+                    __A: _serde::de::MapAccess<'de>,
+                {
+                    let mut __field0: _serde::__private228::Option<String> = _serde::__private228::None;
+                    while let _serde::__private228::Some(__key) = _serde::de::MapAccess::next_key::<
+                        __Field,
+                    >(&mut __map)? {
+                        match __key {
+                            __Field::__field0 => {
+                                if _serde::__private228::Option::is_some(&__field0) {
+                                    return _serde::__private228::Err(
+                                        <__A::Error as _serde::de::Error>::duplicate_field(
+                                            "payload",
+                                        ),
+                                    );
+                                }
+                                __field0 = _serde::__private228::Some(
+                                    _serde::de::MapAccess::next_value::<String>(&mut __map)?,
+                                );
+                            }
+                            _ => {
+                                let _ = _serde::de::MapAccess::next_value::<
+                                    _serde::de::IgnoredAny,
+                                >(&mut __map)?;
+                            }
+                        }
+                    }
+                    let __field0 = match __field0 {
+                        _serde::__private228::Some(__field0) => __field0,
+                        _serde::__private228::None => {
+                            _serde::__private228::de::missing_field("payload")?
+                        }
+                    };
+                    _serde::__private228::Ok(Echo { payload: __field0 })
+                }
+            }
+            #[doc(hidden)]
+            const FIELDS: &'static [&'static str] = &["payload"];
+            _serde::Deserializer::deserialize_struct(
+                __deserializer,
+                "Echo",
+                FIELDS,
+                __Visitor {
+                    marker: _serde::__private228::PhantomData::<Echo>,
+                    lifetime: _serde::__private228::PhantomData,
+                },
+            )
+        }
+    }
+};
+impl ::trillium_api::TryFromConn for Echo {
+    type Error = ::trillium_api::Error;
+    async fn try_from_conn(
+        conn: &mut ::trillium::Conn,
+    ) -> ::core::result::Result<Self, Self::Error> {
+        ::trillium_api::ApiConnExt::deserialize::<Self>(conn).await
+    }
+}
+impl ::trillium::Handler for Echo {
+    async fn run(&self, mut conn: ::trillium::Conn) -> ::trillium::Conn {
+        match ::trillium_api::ApiConnExt::serialize(&mut conn, &*self).await {
+            ::core::result::Result::Ok(()) => conn,
+            ::core::result::Result::Err(e) => conn.with_state(e).halt(),
+        }
+    }
+    async fn before_send(&self, mut conn: ::trillium::Conn) -> ::trillium::Conn {
+        if let ::core::option::Option::Some(error) = conn
+            .take_state::<::trillium_api::Error>()
+        {
+            ::trillium::Handler::before_send(&error, conn).await
+        } else {
+            conn
+        }
+    }
+}