+++ api/src/halt.rs
@@ -2,7 +2,7 @@ use trillium::{Conn, Handler};
 
 /// a struct that halts the Conn handler sequence. see [`Conn::halt`]
 /// for more.
-#[derive(Clone, Copy, Debug)]
+#[derive(Clone, Copy, Debug, Default)]
 pub struct Halt;
 
 impl Handler for Halt {